requirements.txt
google-generativeai
gtts
playsound
python-dotenv